Add mutex to protect connections fixing a race condition
The connections map can be accessed from different goroutines at the same
time. For example handleDisconnectNotification which is modifying the map is
running in a different goroutine than the echo and update handlers which are
accessing the map. This leads in a race of the access to the map
connections. Thus calling the Diconnect func can trigger the same issue.

Running the test suite with the -race show the issue.